+++ pkg/sql/paramparse/BUILD.bazel
@@ -3,29 +3,17 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "paramparse",
     srcs = [
-        "paramobserver.go",
         "paramparse.go",
         "validation.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/paramparse",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/geo/geoindex",
-        "//pkg/server/telemetry",
-        "//pkg/settings",
-        "//pkg/sql/catalog/catpb",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
-        "//pkg/sql/pgwire/pgnotice",
         "//pkg/sql/sem/eval",
-        "//pkg/sql/sem/normalize",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sqltelemetry",
-        "//pkg/sql/types",
         "//pkg/util/duration",
-        "//pkg/util/errorutil/unimplemented",
         "@com_github_cockroachdb_errors//:errors",
     ],
 )
