+++ pkg/sql/storageparam/BUILD.bazel
@@ -1,0 +1,20 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "storageparam",
+    srcs = ["storage_param.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/storageparam",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/server/telemetry",
+        "//pkg/sql/paramparse",
+        "//pkg/sql/pgwire/pgcode",
+        "//pkg/sql/pgwire/pgerror",
+        "//pkg/sql/pgwire/pgnotice",
+        "//pkg/sql/sem/eval",
+        "//pkg/sql/sem/normalize",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/sqltelemetry",
+        "//pkg/sql/types",
+    ],
+)
